Fixed player movement; issue w/ raycasting
I fixed the issues with the movement by simply copying the tutorial code and cutting out things I didn't need. I think what happened is something didn't get written properly when I was following along.

There's now an issue with the raycasts. I know it has to do with the pivot point of the center of the player game object, but I don't know how to move it atm.