Added rudimentary collectable item
The main collectable-- the gemstone-- has been added. Currently, it's a prefab that disappears when the player walks into it. The item deals with gem collection and score augmentation, but the player movement script is what detects when the player walks into the item. Not sure if I'll keep gem detection specifically within the player movement script.

+ Made a bigger arena for the player to explore for the demo