BUG: Player hitbox gets stuck when trying to jump on certain geometry
I have an inkling of an idea as to why this bug might be occurring
What ledge hanging does is it forces the player's hitbox close to/into the ground and then forces it upwards, but if the ledge the player holds onto is of a certain width or forces the player in a certain depth then the hitbox gets stuck inside the ground. As a result the player "jumps" but is forced off the ledge as if they were crouching.

will need to fix later